refactor: split status page components
