+++ src/components/payment-status/EmptyState.tsx
@@ -1,0 +1,7 @@
+export function EmptyState() {
+  return (
+    <div className="text-center py-8">
+      <p className="text-gray-500">No paid orders found</p>
+    </div>
+  );
+}
